+++ admin/reviews.php
@@ -49,13 +49,25 @@ define('myyshop', true);
         switch ($action) {
             case 'accept':
             
-            $update = mysql_query("UPDATE table_reviews SET moderat='1' WHERE reviews_id='$id'",$link);
+            if($_SESSION['accept_reviews'] == '1')
+            {
+                $update = mysql_query("UPDATE table_reviews SET moderat='1' WHERE reviews_id='$id'",$link);
+            }else
+            {
+                $msgerror = 'У вас нет прав на одобрение отзывов';
+            }
             
             break;
             
             case 'delete':
             
-            $delete = mysql_query("DELETE FROM table_reviews WHERE reviews_id='$id'",$link);
+            if($_SESSION['delete_reviews'] == '1')
+            {
+                $delete = mysql_query("DELETE FROM table_reviews WHERE reviews_id='$id'",$link);
+            }else
+            {
+                $msgerror = 'У вас нет прав на удаление отзывов';
+            }           
             
             break;
         }
